refactored admin endpoint functions into separate files
